fix(form-field): async checkbox labels in the template are overwritten
The check for an existing checkbox label didn't account for labels set with an async pipe and also
deleted the text node where angular wanted to set the label text when it was resolved. Now the text
node will be updated instead of overwritten, so angular can revert the label if it is set with an
async pipe.